teamrepoistory int to Long

+++ src/main/java/CRM/repository/TeamRepository.java
@@ -3,6 +3,6 @@ package CRM.repository;
 import CRM.domain.TeamEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeamRepository extends JpaRepository<TeamEntity, Integer> {
+public interface TeamRepository extends JpaRepository<TeamEntity, Long> {
 
 }
